integrate wdt into controller
