+++ Hocki_PIO/.vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "files.associations": {
-        "array": "cpp"
+        "array": "cpp",
+        "*.tpp": "cpp"
     }
 }
