+++ Hocki_PIO/lib/Controller/Controller.h
@@ -4,15 +4,17 @@
 #include "Arduino.h"
 #include "sbus.h"
 #include "ControllerPacket.h"
+#include "Watchdog_t4.h"
 
 class Controller {
-    private:
-      SbusRx *sbus_rx;
-      bool startedReading;
-    public:
-        Controller();
-        void init();
-        bool read(ControllerPacket *packet);
+	private:
+		SbusRx *sbus_rx;
+		bool startedReading;
+	public:
+		Controller();
+		WDT_T4<WDT3> wdt;
+		void init();
+		bool read(ControllerPacket *packet);
 };
 
 #endif
